refine word frequency to include emails
in this fix I have implemented code to search and find emails from the
line read, build a new string with emails only which are separated by
spaces and then replace those emails  with an empty string on the
original line. Then I build a new line by concatinating the original
line without emails with the emails separated space.